backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
statsmodels
python-docx
pypdf
openpyxl
python-multipart
pytest
